organized the components folder so that it's easier to understand. Made other folders inside components folder
