+++ src/components/DashboardFiles/EditNotes.js
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react'
 
 
 // This will be for editing notes to cities in the list
-const Edit = () => {
+const EditNotes = () => {
 
     return (
         <>
@@ -11,4 +11,4 @@ const Edit = () => {
     )
 }
 
-export default Edit;
+export default EditNotes;
